Refactor `useModal` hook and extract utility functions
- Moved `handleSaveFavoritePhoto` and `handleOverlayClick` to separate utility functions
- Created `useFetchFavorites` and `useBodyScrollLock` hooks for better separation of concerns
- Updated `useModal` hook to utilize these utility functions and custom hooks
