+++ src/components/Modal/useBodyScrollLock.js
@@ -1,0 +1,10 @@
+import { useEffect } from 'react';
+
+export const useBodyScrollLock = () => {
+  useEffect(() => {
+    document.body.style.overflow = 'hidden';
+    return () => {
+      document.body.style.overflow = 'auto';
+    };
+  }, []);
+};
