+++ src/pages/Favorites/useFetchFavorites.js
@@ -1,0 +1,11 @@
+import { useEffect } from 'react';
+
+import { getFavorites } from '@/store/favoritesSlice';
+
+export const useFetchFavorites = (user, favorites, dispatch) => {
+  useEffect(() => {
+    if (user && favorites.length === 0) {
+      dispatch(getFavorites());
+    }
+  }, [dispatch, user, favorites.length]);
+};
